+++ unisci/__init__.py
@@ -1,3 +1,3 @@
-__version__ = "1.3.1"
+__version__ = "1.4.0"
 
 from .types import *    # imports Element, Quantity, Temperature